+++ src/files/index.js
@@ -14,6 +14,12 @@ import { toCamelCase } from '../utils/objects';
  */
 
 /**
+ * @typedef {Object} FileToDownload
+ * @property {string} expiresAt ISO 8601 Extended Format date/time
+ * @property {string} temporaryUrl A temporary URL that can be used to download the file
+ */
+
+/**
  * Module that provides access to information about Files
  *
  * @typicalname contxtSdk.files
@@ -29,6 +35,36 @@ class Files {
     this._baseUrl = baseUrl;
     this._request = request;
     this._sdk = sdk;
+  }
+
+  /**
+   * Gets a temporary URL for the file.
+   *
+   * API Endpoint: '/files/:fileId/download'
+   * Method: GET
+   *
+   * @param {string} fileId The ID of the file
+   *
+   * @returns {Promise}
+   * @fulfill {FileToDownload} Information needed to download the file
+   * @reject {Error}
+   *
+   * @example
+   * contxtSdk.files
+   *   .download('bbcdd201-58f7-4b69-a24e-752e9490a347')
+   *   .then((file) => console.log(file))
+   *   .catch((err) => console.log(err));
+   */
+  download(fileId) {
+    if (!fileId) {
+      return Promise.reject(
+        new Error('A file ID is required for downloading a file')
+      );
+    }
+
+    return this._request
+      .get(`${this._baseUrl}/files/${fileId}/download`)
+      .then((file) => toCamelCase(file));
   }
 
   /**