+++ support/fixtures/factories/index.js
@@ -27,6 +27,7 @@ require('./fieldCategory');
 require('./fieldGrouping');
 require('./fieldGroupingField');
 require('./file');
+require('./fileToDownload');
 require('./outputField');
 require('./outputFieldData');
 require('./organization');